requirements.txt
newsapi-python
httpx
python-dotenv
agno
google-genai
